Optionally open in a tab instead of a dialog
This change introduces opening wallabag in a tab, which is also reused
if possible. This is an option that can be configured by the user.

Secondly, I also implemented an option that disables the autoclose
parameter to wallabag. This is problematic when the user configured
wallabag to use http basic auth, as the user cannot see the success of
adding a link to wallabag.