Fix tab reuse, activate reused tab
This change fixes reusing a tab. Previously the onOpen and onClose
methods would react on *all* tab open and close events, which was not
what I wanted. Now I have an onOpen and onClose handler that only listen
to events emited from the tab I created.

I also added that if a tab is already in existence, it is not only
feeded with the new url, but it is also brought into focus.